# nodejs-server-server-generated/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Whats App API"
  version: "1.0.0"
  title: "Whats App API"
  contact:
    name: "Irene-Maria Tabakis"
    email: "irma.tabakis@gmail.com"
host: "virtserver.swaggerhub.com"
basePath: "/asterakia/WhatsApp/1.0.0"
...
schemes:
- "https"
paths:
...
  /media:
    post:
      tags:
      - "media"
      summary: "Upload a file"
      operationId: "mediaPOST"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "file"
        in: "formData"
        description: "The file which is being uploaded"
        required: true
        type: "string"
      responses:
        200:
          description: "Sucessfully uploaded the file. The variable name of the response\
            \ is media."
          schema:
            $ref: "#/definitions/inline_response_200_12"
      security:
      - Authorization: []
      x-swagger-router-controller: "Media"
  /media/shared/{userId1}/{userId2}:
    get:
      tags:
      - "media"
      summary: "Get all media which two user's have shared between each other"
      operationId: "mediaSharedUserId1UserId2GET"
      parameters:
      - name: "userId1"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      - name: "userId2"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "Returns an array of all media two users have shared between\
            \ themeselves"
          schema:
            $ref: "#/definitions/inline_response_200_13"
      security:
      - Authorization: []
      x-swagger-router-controller: "Media"
  /media/{mediaId}:
    delete:
      tags:
      - "media"
      summary: "Delete a media object."
      operationId: "mediaMediaIdDELETE"
      parameters:
      - name: "mediaId"
        in: "path"
        description: "The ID of the specific media object the developer wishes to\
          \ delete"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "The media objected was successfully deleted"
          schema:
            $ref: "#/definitions/Media"
      security:
      - Authorization: []
      x-swagger-router-controller: "Media"
...
  /groupChats:
    post:
      tags:
      - "groupchat"
      summary: "Create a groupChat"
      operationId: "groupChatsPOST"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "userIds"
        in: "formData"
        description: "An array of userIds. Can also be just one userId (in an array\
          \ with the name userIds) if only the admin (creator) is the only user in\
          \ the group when it is created"
        required: true
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      - name: "name"
        in: "formData"
        description: "The name of the group chat"
        required: false
        type: "string"
      - name: "adminId"
        in: "formData"
        description: "The ID of the user who is creating the group chat"
        required: true
        type: "string"
      - name: "mediaId"
        in: "formData"
        description: "If the groupChat has an image assoicated with it, the media\
          \ Id can be used to create that image"
        required: false
        type: "string"
      responses:
        200:
          description: "Returns the chat which was just created. The variable name\
            \ of the response is chat."
          schema:
            $ref: "#/definitions/inline_response_200_15"
      security:
      - Authorization: []
      x-swagger-router-controller: "Groupchat"
  /groupChats/{groupChatId}:
    put:
      tags:
      - "groupchat"
      summary: "Edit the details of a groupChat"
      operationId: "groupChatsGroupChatIdPUT"
      parameters:
      - name: "groupChatId"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/body_6"
      responses:
        200:
          description: "Returns the updated groupChat object on success"
          schema:
            $ref: "#/definitions/inline_response_200_16"
      security:
      - Authorization: []
      x-swagger-router-controller: "Groupchat"
  /groupChats/{chatId}/join:
    put:
      tags:
      - "groupchat"
      summary: "Add user(s) to a groupChat"
      operationId: "groupChatsChatIdJoinPUT"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "userIds"
        in: "formData"
        description: "An array of userIds. Can also be just one userId (in an array\
          \ with the name userIds) if only the admin (creator) is the only user in\
          \ the group when it is created"
        required: true
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      - name: "chatId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "Returns the chat object in which the user(s) just joined.\
            \ The variable name of the response is chat."
          schema:
            $ref: "#/definitions/Chat"
      security:
      - Authorization: []
      x-swagger-router-controller: "Groupchat"
  /groupChats/{chatId}/user/{userId}/messages:
    get:
      tags:
      - "groupchat"
      summary: "Get 40 messages belonging to a group chat"
      operationId: "groupChatsChatIdUserUserIdMessagesGET"
      parameters:
      - name: "chatId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      - name: "userId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      - name: "messageId"
        in: "query"
        description: "Include this query parameter when you would like to get the\
          \ next 40 messages following a specific message. If this query parameter\
          \ is not included, you will get the 40 newest messages in order"
        required: false
        type: "string"
      responses:
        200:
          description: "Returns all the messages associated with a specific group\
            \ chat. The messages will be divided into two sections -- unread & read.\
            \ In each section, the messages will be ordered from newest to oldest.\
            \ The variable name of the response is messages. allMessages will be ordered\
            \ from oldest to newest (for you Sung!). The client can figure out which\
            \ messages are unread by comparing the createdAt attribute with the last_time_user_entered\
            \ variable. This route will trigger the \"read-receipts\" socket function\
            \ for the specific user querying for message history"
          schema:
            $ref: "#/definitions/inline_response_200_17"
      security:
      - Authorization: []
      x-swagger-router-controller: "Groupchat"
  /groupChats/{chatId}/favorite:
    put:
      tags:
      - "groupchat"
      summary: "Favorite a specific group chat"
      operationId: "groupChatsChatIdFavoritePUT"
      parameters:
      - name: "chatId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/body_7"
      responses:
        200:
          description: "Returns the chat which was just favorited. The variable name\
            \ of the response is chat."
          schema:
            $ref: "#/definitions/Chat"
      security:
      - Authorization: []
      x-swagger-router-controller: "Groupchat"
...
  /chat/{chatId}:
    get:
      tags:
      - "chat"
      summary: "Get a chat by ID"
      operationId: "chatChatIdGET"
      parameters:
      - name: "chatId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "Returns a chat object and the response variable name is also\
            \ chat"
          schema:
            $ref: "#/definitions/Chat"
      security:
      - Authorization: []
      x-swagger-router-controller: "Chat"
  /chat/{chatId}/file:
    get:
      tags:
      - "chat"
      summary: "Get all files belonging to a specific chat"
      operationId: "chatChatIdFileGET"
      parameters:
      - name: "chatId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "Returns the files which were shared in a chat"
          schema:
            $ref: "#/definitions/inline_response_200_19"
      security:
      - Authorization: []
      x-swagger-router-controller: "Chat"
  /chat/{chatId}/media:
    get:
      tags:
      - "chat"
      summary: "Returns all the media files shared in a chat"
      operationId: "chatChatIdMediaGET"
      parameters:
      - name: "chatId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "Returns all the media files shared in a chat"
          schema:
            $ref: "#/definitions/inline_response_200_20"
      security:
      - Authorization: []
      x-swagger-router-controller: "Chat"
  /chat/{chatId}/user/{userId}/notifications:
    put:
      tags:
      - "chat"
      summary: "Edit the notifications a user can receive from a chat"
      operationId: "chatChatIdUserUserIdNotificationsPUT"
      parameters:
      - name: "chatId"
        in: "path"
        required: true
        type: "string"
      - name: "userId"
        in: "path"
        required: true
        type: "string"
      - name: "user_added_notifications"
        in: "query"
        required: false
        type: "boolean"
      - name: "message_notifications"
        in: "query"
        required: false
        type: "boolean"
      responses:
        200:
          description: "Returns the update userChat which includes the attribitues\
            \ message_notifications and user_added_notifications which indicated the\
            \ specific nofitifcation options for a specifc chat"
          schema:
            $ref: "#/definitions/inline_response_200_21"
      security:
      - Authorization: []
      x-swagger-router-controller: "Chat"
...
  /call/starts:
    post:
      tags:
      - "call"
      summary: "New call starts."
      operationId: "callStartsPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "post"
        in: "query"
        description: "If post is set to true, a call starts that needs to be added\
          \ to the DB."
        required: true
        type: "boolean"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/body_11"
      responses:
        200:
          description: "On success, the messageInfo object will be sent to the socket\
            \ functions. The variable name of the response is messageInfo"
          schema:
            $ref: "#/definitions/Message"
        400:
          description: "This error can occur if the chatId or userID does not exist,\
            \ if the user is not a participant in the chat, or if the two users have\
            \ a relationship of blocked"
      security:
      - Authorization: []
      - Socket: []
      x-swagger-router-controller: "Call"
  /call/ends:
    post:
      tags:
      - "call"
      summary: "New call starts."
      operationId: "callEndsPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "post"
        in: "query"
        description: "If post is set to true, a call starts that needs to be added\
          \ to the DB."
        required: true
        type: "boolean"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/body_12"
      responses:
        200:
          description: "On success, the messageInfo object will be sent to the socketfunctions.\
            \ The variable name of the response is messageInfo"
          schema:
            $ref: "#/definitions/Message"
        400:
          description: "This error can occur if the chatId or userID does not exist,\
            \ if the user is not a participant in the chat, or if the two users have\
            \ a relationship of blocked"
      security:
      - Authorization: []
      - Socket: []
      x-swagger-router-controller: "Call"
...
securityDefinitions:
  Authorization:
    type: "apiKey"
    name: "Authorization"
    in: "header"
  Socket:
    type: "apiKey"
    name: "socket"
    in: "header"
definitions:
...
  Media:
    type: "object"
    properties:
      id:
        type: "number"
      media_file_string:
        type: "string"
        description: "In the media table, there is a getter function which retrieves\
          \ the AWS bucket file which this media object is assoicated with"
      media_file:
        type: "string"
        description: "The URL which the file can be uploaded from"
      user_id:
        type: "number"
      groupChatId:
        type: "number"
    example:
      media_file_string: "media_file_string"
      groupChatId: 1.46581298050294517310021547018550336360931396484375
      user_id: 6.02745618307040320615897144307382404804229736328125
      id: 0.80082819046101150206595775671303272247314453125
      media_file: "media_file"
  Chat:
    type: "object"
    properties:
      id:
        type: "number"
      groupChat:
        type: "boolean"
        description: "Set to true if the chat is a group chat"
      privateChat:
        type: "boolean"
        description: "Set tot true if the chat is a private chat"
      participantsId:
        type: "array"
        example:
        - "22"
        - "33"
        description: "An array of the user Ids participating in the chat"
        items:
          type: "integer"
      last_activity:
        type: "string"
      last_activity_in_associated_channel_event_or_groupChat:
        type: "string"
    example:
      last_activity: "last_activity"
      groupChat: true
      privateChat: true
      id: 0.80082819046101150206595775671303272247314453125
      participantsId:
      - "22"
      - "33"
      last_activity_in_associated_channel_event_or_groupChat: "last_activity_in_associated_channel_event_or_groupChat"
  UserChats:
    type: "object"
    properties:
      userId:
        type: "string"
      chatId:
        type: "string"
      favorite:
        type: "boolean"
      last_time_in_chat:
        type: "string"
      unread_messages:
        type: "boolean"
      entered_chat_quantity:
        type: "number"
      current_chat:
        type: "boolean"
      message_notifications:
        type: "boolean"
      user_added_notifications:
        type: "boolean"
      messages_not_to_display:
        type: "array"
        example:
        - 50
        - 100
        - 150
        items:
          type: "number"
          description: "The message ID's which will not be displayed to a user because\
            \ they were self-destructing messages"
    example:
      last_time_in_chat: "last_time_in_chat"
      chatId: "chatId"
      entered_chat_quantity: 0.80082819046101150206595775671303272247314453125
      current_chat: true
      messages_not_to_display:
      - 50
      - 100
      - 150
      unread_messages: true
      user_added_notifications: true
      userId: "userId"
      favorite: true
      message_notifications: true
  Message:
    type: "object"
    properties:
      id:
        type: "number"
      message:
        type: "string"
      chatId:
        type: "string"
      userId:
        type: "string"
      type:
        type: "string"
        example: "media, string"
      createdAt:
        type: "string"
        description: "The date time of when the message was created (sent)"
      name:
        type: "string"
        description: "The full name of the user who sent the message"
      post:
        type: "boolean"
      user:
        $ref: "#/definitions/inline_response_200_7_message_user"
      GroupChat:
        $ref: "#/definitions/Message_GroupChat"
      mediaArray:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_7_message_mediaArray"
      files:
        type: "array"
        items:
          $ref: "#/definitions/Message_files"
    example:
      createdAt: "createdAt"
      GroupChat:
        socket_id: "socket_id"
        verified: true
        last_name: "last_name"
        avatar: "avatar"
        profile_picture_string: "profile_picture_string"
        country_code: "country_code"
        passcode_expiration: "passcode_expiration"
        phone: "phone"
        id: 6.02745618307040320615897144307382404804229736328125
        position: "{}"
        first_name: "first_name"
        email: "email"
        passcode: "passcode"
      chatId: "chatId"
      post: true
      name: "name"
      files:
      - s3_url: "s3_url"
        file_string: "file_string"
      - s3_url: "s3_url"
        file_string: "file_string"
      id: 0.80082819046101150206595775671303272247314453125
      message: "message"
      type: "media, string"
      userId: "userId"
      user:
        avatar: "avatar"
        profile_picture_string: "profile_picture_string"
      mediaArray:
      - media_file_string: "media_file_string"
        media_file: "media_file"
      - media_file_string: "media_file_string"
        media_file: "media_file"
  GroupChat:
    type: "object"
    properties:
      id:
        type: "integer"
      chatId:
        type: "integer"
      name:
        type: "string"
      adminId:
        type: "string"
        description: "The admin is the creator of the group chat. This individual\
          \ has the power to kick users out of the group chat as well"
      createdAt:
        type: "string"
      profile_picture_string:
        type: "string"
      avatar:
        type: "string"
        description: "The url which links to amazon s3 for the groupChat profile image"
    example:
      createdAt: "createdAt"
      chatId: 6
      name: "name"
      adminId: "adminId"
      id: 0
      avatar: "avatar"
      profile_picture_string: "profile_picture_string"
...
  inline_response_200_7_message_user:
    properties:
      avatar:
        type: "string"
      profile_picture_string:
        type: "string"
    example:
      avatar: "avatar"
      profile_picture_string: "profile_picture_string"
  inline_response_200_7_message_mediaArray:
    properties:
      media_file_string:
        type: "string"
      media_file:
        type: "string"
        description: "The url of the link to the file"
    example:
      media_file_string: "media_file_string"
      media_file: "media_file"
...
  inline_response_200_12_mediaArray:
    properties:
      id:
        type: "number"
      media_file_string:
        type: "string"
        description: " "
      media_file:
        type: "string"
      user_id:
        type: "number"
      groupChatId:
        type: "number"
    example:
      media_file_string: "media_file_string"
      groupChatId: 1.46581298050294517310021547018550336360931396484375
      user_id: 6.02745618307040320615897144307382404804229736328125
      id: 0.80082819046101150206595775671303272247314453125
      media_file: "media_file"
  inline_response_200_12:
    properties:
      mediaArray:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_12_mediaArray"
    example:
      mediaArray:
      - media_file_string: "media_file_string"
        groupChatId: 1.46581298050294517310021547018550336360931396484375
        user_id: 6.02745618307040320615897144307382404804229736328125
        id: 0.80082819046101150206595775671303272247314453125
        media_file: "media_file"
      - media_file_string: "media_file_string"
        groupChatId: 1.46581298050294517310021547018550336360931396484375
        user_id: 6.02745618307040320615897144307382404804229736328125
        id: 0.80082819046101150206595775671303272247314453125
        media_file: "media_file"
  inline_response_200_13_sharedMedia:
    properties:
      id:
        type: "number"
      media_file_string:
        type: "string"
        description: "In the media table, there is a getter function which retrieves\
          \ the AWS bucket file which this media object is assoicated with"
      media_file:
        type: "string"
        description: "The URL which the file can be uploaded from"
      user_id:
        type: "number"
      groupChatId:
        type: "number"
    example:
      media_file_string: "media_file_string"
      groupChatId: 1.46581298050294517310021547018550336360931396484375
      user_id: 6.02745618307040320615897144307382404804229736328125
      id: 0.80082819046101150206595775671303272247314453125
      media_file: "media_file"
  inline_response_200_13:
    properties:
      sharedMedia:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_13_sharedMedia"
    example:
      sharedMedia:
      - media_file_string: "media_file_string"
        groupChatId: 1.46581298050294517310021547018550336360931396484375
        user_id: 6.02745618307040320615897144307382404804229736328125
        id: 0.80082819046101150206595775671303272247314453125
        media_file: "media_file"
      - media_file_string: "media_file_string"
        groupChatId: 1.46581298050294517310021547018550336360931396484375
        user_id: 6.02745618307040320615897144307382404804229736328125
        id: 0.80082819046101150206595775671303272247314453125
        media_file: "media_file"
...
  inline_response_200_15_chat:
    properties:
      id:
        type: "number"
      groupChat:
        type: "boolean"
        description: "Set to true if the chat is a group chat"
      privateChat:
        type: "boolean"
        description: "Set tot true if the chat is a private chat"
      participantsId:
        type: "array"
        example:
        - "22"
        - "33"
        description: "An array of the user Ids participating in the chat"
        items:
          type: "integer"
    example:
      groupChat: true
      privateChat: true
      id: 0.80082819046101150206595775671303272247314453125
      participantsId:
      - "22"
      - "33"
  inline_response_200_15:
    properties:
      chat:
        $ref: "#/definitions/inline_response_200_15_chat"
    example:
      chat:
        groupChat: true
        privateChat: true
        id: 0.80082819046101150206595775671303272247314453125
        participantsId:
        - "22"
        - "33"
  body_6:
    type: "object"
    properties:
      name:
        type: "string"
      mediaId:
        type: "string"
  inline_response_200_16:
    properties:
      success:
        type: "boolean"
      groupChat:
        $ref: "#/definitions/GroupChat"
    example:
      groupChat:
        createdAt: "createdAt"
        chatId: 6
        name: "name"
        adminId: "adminId"
        id: 0
        avatar: "avatar"
        profile_picture_string: "profile_picture_string"
      success: true
  inline_response_200_17_messages_files:
    properties:
      url:
        type: "string"
      file_string:
        type: "string"
    example:
      file_string: "file_string"
      url: "url"
...
  inline_response_200_17:
    properties:
      messages:
        $ref: "#/definitions/inline_response_200_17_messages"
    example:
      messages:
        read:
        - createdAt: "createdAt"
          chatId: 2.3021358869347654518833223846741020679473876953125
          id: 5.962133916683182377482808078639209270477294921875
          message: "message"
          type: "type"
          userId: 5.63737665663332876420099637471139430999755859375
          mediaArray:
          - "mediaArray"
          - "mediaArray"
          user:
            avatar: "avatar"
            profile_picture_string: "profile_picture_string"
        - createdAt: "createdAt"
          chatId: 2.3021358869347654518833223846741020679473876953125
          id: 5.962133916683182377482808078639209270477294921875
          message: "message"
          type: "type"
          userId: 5.63737665663332876420099637471139430999755859375
          mediaArray:
          - "mediaArray"
          - "mediaArray"
          user:
            avatar: "avatar"
            profile_picture_string: "profile_picture_string"
        last_time_user_entered: "last_time_user_entered"
        unread:
        - createdAt: "createdAt"
          chatId: 1.46581298050294517310021547018550336360931396484375
          files:
          - file_string: "file_string"
            url: "url"
          - file_string: "file_string"
            url: "url"
          id: 0.80082819046101150206595775671303272247314453125
          message: "message"
          type: "type"
          userId: 6.02745618307040320615897144307382404804229736328125
          mediaArray:
          - "mediaArray"
          - "mediaArray"
          user:
            avatar: "avatar"
            profile_picture_string: "profile_picture_string"
        - createdAt: "createdAt"
          chatId: 1.46581298050294517310021547018550336360931396484375
          files:
          - file_string: "file_string"
            url: "url"
          - file_string: "file_string"
            url: "url"
          id: 0.80082819046101150206595775671303272247314453125
          message: "message"
          type: "type"
          userId: 6.02745618307040320615897144307382404804229736328125
          mediaArray:
          - "mediaArray"
          - "mediaArray"
          user:
            avatar: "avatar"
            profile_picture_string: "profile_picture_string"
        allMessages:
        - createdAt: "createdAt"
          chatId: 2.3021358869347654518833223846741020679473876953125
          id: 5.962133916683182377482808078639209270477294921875
          message: "message"
          type: "type"
          userId: 5.63737665663332876420099637471139430999755859375
          mediaArray:
          - "mediaArray"
          - "mediaArray"
          user:
            avatar: "avatar"
            profile_picture_string: "profile_picture_string"
        - createdAt: "createdAt"
          chatId: 2.3021358869347654518833223846741020679473876953125
          id: 5.962133916683182377482808078639209270477294921875
          message: "message"
          type: "type"
          userId: 5.63737665663332876420099637471139430999755859375
          mediaArray:
          - "mediaArray"
          - "mediaArray"
          user:
            avatar: "avatar"
            profile_picture_string: "profile_picture_string"
  body_7:
    type: "object"
    properties:
      userId:
        type: "string"
      trueOrFalse:
        type: "boolean"
        description: "Set to true, if you would like to make the chat a favorite.\
          \ Set to false if you would like to unfavorite a chat"
...
  inline_response_200_19:
    properties:
      success:
        type: "boolean"
      files:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_17_messages_files"
    example:
      success: true
      files:
      - file_string: "file_string"
        url: "url"
      - file_string: "file_string"
        url: "url"
  inline_response_200_20_mediaArray:
    properties:
      media_file:
        type: "string"
      media_file_string:
        type: "string"
    example:
      media_file_string: "media_file_string"
      media_file: "media_file"
  inline_response_200_20:
    properties:
      success:
        type: "boolean"
      mediaArray:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_20_mediaArray"
    example:
      success: true
      mediaArray:
      - media_file_string: "media_file_string"
        media_file: "media_file"
      - media_file_string: "media_file_string"
        media_file: "media_file"
  inline_response_200_21:
    properties:
      success:
        type: "boolean"
      userChat:
        $ref: "#/definitions/UserChats"
    example:
      userChat:
        last_time_in_chat: "last_time_in_chat"
        chatId: "chatId"
        entered_chat_quantity: 0.80082819046101150206595775671303272247314453125
        current_chat: true
        messages_not_to_display:
        - 50
        - 100
        - 150
        unread_messages: true
        user_added_notifications: true
        userId: "userId"
        favorite: true
        message_notifications: true
      success: true
...
  body_11:
    type: "object"
    properties:
      userId:
        type: "number"
      call:
        type: "string"
      chatId:
        type: "number"
      createdAt:
        type: "string"
        description: "The createdAt attribute is the date time a call was created\
          \ (send)."
  body_12:
    type: "object"
    properties:
      userId:
        type: "string"
      call:
        type: "string"
      chatId:
        type: "string"
      finishedAt:
        type: "string"
        description: "The finishedAt attribute is the date time a call was finished\
          \ (end)."
...
  Message_files:
    properties:
      file_string:
        type: "string"
      s3_url:
        type: "string"
    example:
      s3_url: "s3_url"
      file_string: "file_string"